[util] add log initialize and signal handler
We default to printing all logs to stderr, which is the
current behaviour when abseil::log is not initialized.
Only add the initialization to the play target for now. Any new
executables will need to initialize logging as a part of startup.

This commit also enables the abseil signal handler. This attempts to
print a stack trace on signals such as sigsev.
This code is not documented in the online abseil docs
but is in the public interface here:
https://github.com/abseil/abseil-cpp/blob/master/absl/debugging/failure_signal_handler.h